% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellVit.R
\name{CellVit}
\alias{CellVit}
\title{Compute the cell vitality}
\usage{
CellVit(unstained, stained)
}
\arguments{
\item{unstained}{cell count of unstained cells}

\item{stained}{cell count of stained cells}

\item{unstained_prefro}{in adavance of frosting}

\item{unstained_defro}{after frosting}

\item{unstained_24h}{24h after defrosting (and cultivation)}
}
\value{
vitality
}
\description{
Compute the cell vitality
}
