% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absorption_spectroscopy.R
\name{conc.from.abs}
\alias{conc.from.abs}
\title{Functions to compute fundamental values in absorption-spectroscopy}
\usage{
conc.from.abs(abs, epsilon, cuvette = 1, dfac = 1)
}
\description{
Functions to compute fundamental values in absorption-spectroscopy
}
