% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dilfact.R
\name{dilfact}
\alias{dilfact}
\title{Computing a dilution factor, a sample volume (Vorlagevolumen), concentration after dilution}
\usage{
dilfact(vol.1, vol.2)
}
\description{
Computing a dilution factor, a sample volume (Vorlagevolumen), concentration after dilution
}
