% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eadie_Hofstee.R
\name{Eadie_Hofstee}
\alias{Eadie_Hofstee}
\title{Draw an Eadie-Hofstee graph (and compute the y-axis intercept)}
\usage{
Eadie_Hofstee(vel, sub)
}
\arguments{
\item{vel}{Velocity}

\item{sub}{Substarate concentration}
}
\description{
Draw an Eadie-Hofstee graph (and compute the y-axis intercept)
}
