% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conc_eval.R
\name{Konzentration}
\alias{Konzentration}
\title{Generischer Workflow Konzentrationsbestimmung durch Regression}
\usage{
Konzentration(abs_P)
}
\arguments{
\item{abs_P}{float}
}
\value{
float
}
\description{
Hiermit kannst du die Ergebnisse deiner Standardreihe (Absrptionsmessung jeder Art) Plotten, bewerten und
die Konzentration deinrt Unbekannten Probe bestimmen.
Verdünnunsfaktoren sind noch kein Bestandteil der Funktionen
}
