% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lineweaver_Burk.R
\name{Lineweaver_Burk}
\alias{Lineweaver_Burk}
\title{Plot a Lineweaver-Burk diagram and compute the ordinate intercept}
\usage{
Lineweaver_Burk(sub, vel)
}
\arguments{
\item{sub}{list}

\item{vel}{list}
}
\description{
Plot a Lineweaver-Burk diagram and compute the ordinate intercept
}
