% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model.R
\name{LinMod_CEv}
\alias{LinMod_CEv}
\title{Linear model for concentration evaluation}
\usage{
LinMod_CEv(abs, conc)
}
\arguments{
\item{abs}{list}

\item{conc}{list}
}
\value{
tataframe
}
\description{
This must be performed in advanvce of usage of the function conc_eval to produce a linear model suiteable
for concentration evaluation.
}
