% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlrMM.R
\docType{data}
\name{MM_typ}
\alias{MM_typ}
\title{Plot a non linear regression for Michaelis-Menten type enzyme}
\format{An object of class \code{formula} of length 3.}
\usage{
MM_typ
}
\description{
Plot a non linear regression for Michaelis-Menten type enzyme
}
\keyword{datasets}
