% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_linear_MM.R
\name{non_linear_MM}
\alias{non_linear_MM}
\title{Create a non-linear Regression for a Michaelis-Menten type Enzyme with a little knowlege of its Km and Vmax}
\usage{
non_linear_MM(sub, vel, V_max, K_m)
}
\arguments{
\item{sub}{list}

\item{vel}{list}

\item{V_max}{float}

\item{K_m}{float}
}
\description{
Create a non-linear Regression for a Michaelis-Menten type Enzyme with a little knowlege of its Km and Vmax
}
