% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_MM_direkt_1.R
\name{plot_MM_direct}
\alias{plot_MM_direct}
\title{Create a non-linear Regression for a Michaelis-Menten type Enzyme with a little knowlege of its Km and Vmax}
\usage{
plot_MM_direct(sub, velo)
}
\arguments{
\item{sub}{list}

\item{vel}{list}

\item{V_max_start}{float}

\item{K_m_start}{float}
}
\description{
Create a non-linear Regression for a Michaelis-Menten type Enzyme with a little knowlege of its Km and Vmax
}
