% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eadie_Hofstee.R
\name{Eadie_Hofstee}
\alias{Eadie_Hofstee}
\title{Draw an Eadie-Hofstee graph (and compute the y-axis intercept)}
\usage{
Eadie_Hofstee(
  vel,
  sub,
  titleEDH = "Eadie-Hostee-Plot",
  xlable = "vel/sub",
  ylable = "sub"
)
}
\arguments{
\item{vel}{Velocity}

\item{sub}{Substarate concentration}

\item{titleEDH}{Title of the plot}

\item{xlable}{lable of the abscissa}

\item{ylable}{lable of the ordinate}
}
\description{
Draw an Eadie-Hofstee graph (and compute the y-axis intercept)
}
